+++ src/audiovideocontroller/AudioVideoController.ts
@@ -109,5 +109,5 @@ export default interface AudioVideoController extends AudioVideoControllerFacade
   /**
    * Returns the [[EventController]] for this audio-video controller.
    */
-  readonly eventController: EventController;
+  readonly eventController?: EventController;
 }